+++ server/modules/routing/binlog/blr.c
@@ -100,7 +100,7 @@ static  void    errorReply(ROUTER  *instance,
                            error_action_t     action,
                            bool    *succp);
 
-static  int getCapabilities();
+static uint64_t getCapabilities();
 static int blr_handler_config(void *userdata, const char *section, const char *name, const char *value);
 static int blr_handle_config_item(const char *name, const char *value, ROUTER_INSTANCE *inst);
 static int blr_set_service_mysql_user(SERVICE *service);
@@ -1788,9 +1788,9 @@ static void rses_end_locked_router_action(ROUTER_SLAVE *rses)
 }
 
 
-static int getCapabilities()
+static uint64_t getCapabilities()
 {
-    return (int)(RCAP_TYPE_NO_RSESSION | RCAP_TYPE_NO_USERS_INIT);
+    return (RCAP_TYPE_NO_RSESSION | RCAP_TYPE_NO_USERS_INIT);
 }
 
 /**
