+++ server/modules/routing/cli/cli.c
@@ -58,7 +58,7 @@ static  void   closeSession(ROUTER *instance, void *router_session);
 static  void   freeSession(ROUTER *instance, void *router_session);
 static  int    execute(ROUTER *instance, void *router_session, GWBUF *queue);
 static  void   diagnostics(ROUTER *instance, DCB *dcb);
-static  int    getCapabilities();
+static  uint64_t getCapabilities();
 
 /** The module object definition */
 static ROUTER_OBJECT MyObject =
@@ -302,7 +302,7 @@ diagnostics(ROUTER *instance, DCB *dcb)
     return; /* Nothing to do currently */
 }
 
-static int getCapabilities()
+static uint64_t getCapabilities()
 {
     return 0;
 }
