+++ server/modules/routing/maxinfo/maxinfo.c
@@ -79,7 +79,7 @@ static  void    closeSession(ROUTER *instance, void *router_session);
 static  void    freeSession(ROUTER *instance, void *router_session);
 static  int     execute(ROUTER *instance, void *router_session, GWBUF *queue);
 static  void    diagnostics(ROUTER *instance, DCB *dcb);
-static  int     getCapabilities();
+static  uint64_t getCapabilities();
 static  void    handleError(ROUTER         *instance,
                             void           *router_session,
                             GWBUF          *errbuf,
@@ -409,7 +409,7 @@ diagnostics(ROUTER *instance, DCB *dcb)
  *
  * Not used for the maxinfo router
  */
-static int
+static uint64_t
 getCapabilities()
 {
     return 0;
