+++ server/modules/routing/webserver/webserver.c
@@ -51,7 +51,7 @@ static  void    closeSession(ROUTER *instance, void *session);
 static  void    freeSession(ROUTER *instance, void *session);
 static  int     routeQuery(ROUTER *instance, void *session, GWBUF *queue);
 static  void    diagnostic(ROUTER *instance, DCB *dcb);
-static  uint8_t getCapabilities(ROUTER* inst, void* router_session);
+static  uint64_t getCapabilities(ROUTER* inst, void* router_session);
 
 
 static ROUTER_OBJECT MyObject =
@@ -257,7 +257,7 @@ diagnostic(ROUTER *instance, DCB *dcb)
  * @param router_session    The router session
  * @return Router capabilities bitmask
  */
-static uint8_t
+static uint64_t
 getCapabilities(ROUTER *inst, void *router_session)
 {
     return 0;
